fix window not cleaning up properly
also fix warnings in vertex.hpp/vertex.cpp

+++ galaxy/src/galaxy/graphics/Vertex.cpp
@@ -80,15 +80,15 @@ namespace galaxy
 			return *this;
 		}
 
-		std::array<Vertex, 4> Vertex::gen_quad_vertices(const float width, const float height) noexcept
+		std::array<Vertex, 4> Vertex::gen_quad_vertices(const int width, const int height) noexcept
 		{
 			// clang-format off
 			std::array<graphics::Vertex, 4> vertices = 
 			{
 				graphics::Vertex {glm::vec2 {0.0f, 0.0f}, glm::vec2 {0.0f, 0.0f}},
-				graphics::Vertex {glm::vec2 {width, 0.0f}, glm::vec2 {1.0f, 0.0f}},
-				graphics::Vertex {glm::vec2 {width, height}, glm::vec2 {1.0f, 1.0f}},
-				graphics::Vertex {glm::vec2 {0.0f, height}, glm::vec2 {0.0f, 1.0f}}
+				graphics::Vertex {glm::vec2 {static_cast<float>(width), 0.0f}, glm::vec2 {1.0f, 0.0f}},
+				graphics::Vertex {glm::vec2 {static_cast<float>(width), static_cast<float>(height)}, glm::vec2 {1.0f, 1.0f}},
+				graphics::Vertex {glm::vec2 {0.0f, static_cast<float>(height)}, glm::vec2 {0.0f, 1.0f}}
 			};
 			// clang-format on
 
@@ -97,7 +97,7 @@ namespace galaxy
 
 		const std::array<unsigned int, 6>& Vertex::get_default_indices() noexcept
 		{
-			static const std::array<unsigned int, 6> indices = {0, 1, 3, 1, 2, 3};
+			static const std::array<unsigned int, 6> indices = {0u, 1u, 3u, 1u, 2u, 3u};
 			return indices;
 		}
 	} // namespace graphics
